Fixed test_diagnostics error, updated scratch directory in point match residual functions, added option to not delete collection for system_solve, fixed point match reading from file for solve_montage_SL and add 2 node script for generating montage scape point matches.

+++ solver/system_solve.m
@@ -99,7 +99,7 @@ else
     if ~isfield(opts, 'filter_point_matches'), opts.filter_point_matches = 0;end
     if ~isfield(opts, 'use_peg'), opts.use_peg = 0;end
     if ~isfield(opts, 'nbrs_step'), opts.nbrs_step = 1;end
-        
+    if ~isfield(opts, 'delete_existing_collection'), opts.delete_existing_collection = 1; end    
     
     err = [];
     R = [];
@@ -479,7 +479,7 @@ else
         disp('... export to MET (in preparation to be ingested into the Renderer database)...');
         
         v = 'v1';
-        if stack_exists(rcout)
+        if stack_exists(rcout) && opts.delete_existing_collection
             disp('.... removing existing collection');
             resp = delete_renderer_stack(rcout);
         end
